requirements.txt
certifi==2018.1.18
chardet==3.0.4
cmsplugin-filer==1.1.3
Django==1.9
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-cors-headers==2.2.0
django-cors-middleware==1.3.1
django-crum==0.7.2
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-mysql==2.2.2
django-polymorphic==1.3
django-sekizai==0.10.0
Django-Select2==4.3.2
django-treebeard==4.2.1
django-widget-tweaks==1.4.1
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.3
djangocms-video==2.0.4
djangorestframework==3.4.6
easy-thumbnails==2.5
gunicorn==19.7.1
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
jsonfield==2.0.2
mysqlclient==1.3.12
Pillow==5.0.0
PyPDF2==1.26.0
pytz==2018.3
reportlab==3.4.0
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
xhtml2pdf==0.1b2
